rename => GlobalFooter, GlobalNavbar
