+++ src/components/core/Page/GlobalFooter.tsx
